+++ FirebaseAdmin/FirebaseAdmin.Tests/Messaging/FirebaseMessagingClientTest.cs
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 using System;
+using System.Linq;
 using System.Net;
-using System.Threading;
 using System.Threading.Tasks;
 using FirebaseAdmin.Tests;
 using Google.Apis.Auth.OAuth2;
@@ -60,7 +60,7 @@ namespace FirebaseAdmin.Messaging.Tests
         {
             var handler = new MockMessageHandler()
             {
-                Response = new FirebaseMessagingClient.SendResponse()
+                Response = new FirebaseMessagingClient.SingleMessageResponse()
                 {
                     Name = "test-response",
                 },
@@ -71,22 +71,210 @@ namespace FirebaseAdmin.Messaging.Tests
             {
                 Topic = "test-topic",
             };
+
             var response = await client.SendAsync(message);
+
             Assert.Equal("test-response", response);
-            var req = JsonConvert.DeserializeObject<FirebaseMessagingClient.SendRequest>(
-                handler.Request);
+            var req = JsonConvert.DeserializeObject<FirebaseMessagingClient.SendRequest>(handler.Request);
             Assert.Equal("test-topic", req.Message.Topic);
             Assert.False(req.ValidateOnly);
             Assert.Equal(1, handler.Calls);
+            var versionHeader = handler.RequestHeaders.GetValues("X-Firebase-Client").First();
+            Assert.Equal(FirebaseMessagingClient.ClientVersion, versionHeader);
+        }
 
-            // Send in dryRun mode.
-            response = await client.SendAsync(message, dryRun: true);
+        [Fact]
+        public async Task SendDryRunAsync()
+        {
+            var handler = new MockMessageHandler()
+            {
+                Response = new FirebaseMessagingClient.SingleMessageResponse()
+                {
+                    Name = "test-response",
+                },
+            };
+            var factory = new MockHttpClientFactory(handler);
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+            var message = new Message()
+            {
+                Topic = "test-topic",
+            };
+
+            var response = await client.SendAsync(message, dryRun: true);
+
             Assert.Equal("test-response", response);
-            req = JsonConvert.DeserializeObject<FirebaseMessagingClient.SendRequest>(
-                handler.Request);
+            var req = JsonConvert.DeserializeObject<FirebaseMessagingClient.SendRequest>(handler.Request);
             Assert.Equal("test-topic", req.Message.Topic);
             Assert.True(req.ValidateOnly);
-            Assert.Equal(2, handler.Calls);
+            Assert.Equal(1, handler.Calls);
+            var versionHeader = handler.RequestHeaders.GetValues("X-Firebase-Client").First();
+            Assert.Equal(FirebaseMessagingClient.ClientVersion, versionHeader);
+        }
+
+        [Fact]
+        public async Task SendAllAsync()
+        {
+            var rawResponse = @"
+--batch_test-boundary
+Content-Type: application/http
+Content-ID: response-
+
+HTTP/1.1 200 OK
+Content-Type: application/json; charset=UTF-8
+Vary: Origin
+Vary: X-Origin
+Vary: Referer
+
+{
+  ""name"": ""projects/fir-adminintegrationtests/messages/8580920590356323124""
+}
+
+--batch_test-boundary
+Content-Type: application/http
+Content-ID: response-
+
+HTTP/1.1 200 OK
+Content-Type: application/json; charset=UTF-8
+Vary: Origin
+Vary: X-Origin
+Vary: Referer
+
+{
+  ""name"": ""projects/fir-adminintegrationtests/messages/5903525881088369386""
+}
+
+--batch_test-boundary
+";
+            var handler = new MockMessageHandler()
+            {
+                Response = rawResponse,
+                ApplyContentHeaders = (headers) =>
+                {
+                    headers.Remove("Content-Type");
+                    headers.TryAddWithoutValidation("Content-Type", "multipart/mixed; boundary=batch_test-boundary");
+                },
+            };
+            var factory = new MockHttpClientFactory(handler);
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+            var message1 = new Message()
+            {
+                Token = "test-token1",
+            };
+            var message2 = new Message()
+            {
+                Token = "test-token2",
+            };
+
+            var response = await client.SendAllAsync(new[] { message1, message2 });
+
+            Assert.Equal(2, response.SuccessCount);
+            Assert.Equal("projects/fir-adminintegrationtests/messages/8580920590356323124", response.Responses[0].MessageId);
+            Assert.Equal("projects/fir-adminintegrationtests/messages/5903525881088369386", response.Responses[1].MessageId);
+            Assert.Equal(1, handler.Calls);
+            var versionHeader = $"X-Firebase-Client: {FirebaseMessagingClient.ClientVersion}";
+            Assert.Equal(2, this.CountLinesWithPrefix(handler.Request, versionHeader));
+        }
+
+        [Fact]
+        public async Task SendAllAsyncWithError()
+        {
+            var rawResponse = @"
+--batch_test-boundary
+Content-Type: application/http
+Content-ID: response-
+
+HTTP/1.1 200 OK
+Content-Type: application/json; charset=UTF-8
+Vary: Origin
+Vary: X-Origin
+Vary: Referer
+
+{
+  ""name"": ""projects/fir-adminintegrationtests/messages/8580920590356323124""
+}
+
+--batch_test-boundary
+Content-Type: application/http
+Content-ID: response-
+
+HTTP/1.1 400 Bad Request
+Content-Type: application/json; charset=UTF-8
+Vary: Origin
+Vary: X-Origin
+Vary: Referer
+
+{
+  ""error"": {
+    ""code"": 400,
+    ""message"": ""The registration token is not a valid FCM registration token"",
+    ""errors"": [
+      {
+        ""message"": ""The registration token is not a valid FCM registration token"",
+        ""domain"": ""global"",
+        ""reason"": ""badRequest""
+      }
+    ],
+    ""status"": ""INVALID_ARGUMENT""
+  }
+}
+
+--batch_test-boundary
+";
+            var handler = new MockMessageHandler()
+            {
+                Response = rawResponse,
+                ApplyContentHeaders = (headers) =>
+                {
+                    headers.Remove("Content-Type");
+                    headers.TryAddWithoutValidation("Content-Type", "multipart/mixed; boundary=batch_test-boundary");
+                },
+            };
+            var factory = new MockHttpClientFactory(handler);
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+            var message1 = new Message()
+            {
+                Token = "test-token1",
+            };
+            var message2 = new Message()
+            {
+                Token = "test-token2",
+            };
+
+            var response = await client.SendAllAsync(new[] { message1, message2 });
+
+            Assert.Equal(1, response.SuccessCount);
+            Assert.Equal(1, response.FailureCount);
+            Assert.Equal("projects/fir-adminintegrationtests/messages/8580920590356323124", response.Responses[0].MessageId);
+            Assert.NotNull(response.Responses[1].Exception);
+            Assert.Equal(1, handler.Calls);
+            var versionHeader = $"X-Firebase-Client: {FirebaseMessagingClient.ClientVersion}";
+            Assert.Equal(2, this.CountLinesWithPrefix(handler.Request, versionHeader));
+        }
+
+        [Fact]
+        public async Task SendAllAsyncNullList()
+        {
+            var factory = new MockHttpClientFactory(new MockMessageHandler());
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+
+            await Assert.ThrowsAsync<ArgumentNullException>(() => client.SendAllAsync(null));
+        }
+
+        [Fact]
+        public async Task SendAllAsyncWithNoMessages()
+        {
+            var factory = new MockHttpClientFactory(new MockMessageHandler());
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+            await Assert.ThrowsAsync<ArgumentException>(() => client.SendAllAsync(Enumerable.Empty<Message>()));
+        }
+
+        [Fact]
+        public async Task SendAllAsyncWithTooManyMessages()
+        {
+            var factory = new MockHttpClientFactory(new MockMessageHandler());
+            var client = new FirebaseMessagingClient(factory, MockCredential, "test-project");
+            var messages = Enumerable.Range(0, 101).Select(_ => new Message { Topic = "test-topic" });
+            await Assert.ThrowsAsync<ArgumentException>(() => client.SendAllAsync(messages));
         }
 
         [Fact]
@@ -111,6 +299,11 @@ namespace FirebaseAdmin.Messaging.Tests
             Assert.Equal("test-topic", req.Message.Topic);
             Assert.False(req.ValidateOnly);
             Assert.Equal(1, handler.Calls);
+        }
+
+        private int CountLinesWithPrefix(string body, string linePrefix)
+        {
+            return body.Split('\n').Count((line) => line.StartsWith(linePrefix));
         }
     }
 }