+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Unreleased
 
+- [added] Implemented the `SendAllAsync()` and `SendMulticastAsync()` APIs in
+  the `FirebaseMessaging` class.
+
+# v1.2.1
+
 - [fixed] The `VerifyIdTokenAsync()` function now tolerates a clock skew of up
   to 5 minutes when comparing JWT timestamps.